% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{export_data_for_shiny}
\alias{export_data_for_shiny}
\title{export mass_dataset}
\usage{
export_data_for_shiny(object, format, file_path, file_name)
}
\arguments{
\item{compound_id}{kegg cid.}
}
\value{
A dataframe to get term2gene
}
\description{
match kegg cid 2 pathway.
}
