% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{export_final_data}
\alias{export_final_data}
\title{export data}
\usage{
export_final_data(object, path)
}
\arguments{
\item{object}{mass_dataset object.}

\item{path}{file path}
}
\description{
convert mass_dataset to excel format.
}
