% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_features_with_ms2.R
\name{filter_annotations_massdataset}
\alias{filter_annotations_massdataset}
\title{filter features with ms2}
\usage{
filter_annotations_massdataset(
  object,
  annotate_tbl = NULL,
  method = c("only ms2", "only annotation", "both")
)
}
\arguments{
\item{object}{a mass_dataset class}

\item{annotate_tbl}{filtered annotation table.}

\item{method}{filter method,"only ms2","only annotation","both"}
}
\value{
A ggplot object of missing value.
}
\description{
Remove features without MS2 matches.
}
\references{
tiydMass object massdataset
}
