% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_compound_info}
\alias{get_compound_info}
\title{fetch kegg compound id belongs to corresponding pathway}
\usage{
get_compound_info(compound_id)
}
\arguments{
\item{compound_id}{kegg cid.}
}
\value{
A dataframe to get term2gene
}
\description{
match kegg cid 2 pathway.
}
