% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_kegg_pathway_ont}
\alias{get_kegg_pathway_ont}
\title{fetch kegg pathway via ont}
\usage{
get_kegg_pathway_ont(ont)
}
\arguments{
\item{ont}{species id of KEGG database.}
}
\value{
A dataframe to get term2name.
}
\description{
Get pathway id via KEGG API.
}
