% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{gg_message_plot}
\alias{gg_message_plot}
\title{Create empty plot with message}
\usage{
gg_message_plot(print_messages)
}
\arguments{
\item{print_messages}{messages show in plot}
}
\value{
A ggplot object

Return a logical value indicating whether the input file matches the sample ID in the sample information.
}
\description{
Create empty plot with message
}
