% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/homepage_server.R
\name{homepage_server}
\alias{homepage_server}
\title{Homepage Server Module}
\usage{
homepage_server(id)
}
\arguments{
\item{id}{module id}
}
\value{
module server function
}
\description{
Handles server-side logic for the homepage
}
