% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/homepage_ui.R
\name{homepage_ui}
\alias{homepage_ui}
\title{Homepage UI Module}
\usage{
homepage_ui(id)
}
\arguments{
\item{id}{module id}
}
\value{
UI tagList
}
\description{
Creates the homepage interface for tidymassShiny
}
