% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/massqc_pca_3d.R
\name{massqc_pca_3d}
\alias{massqc_pca_3d}
\title{3d pca plot}
\usage{
massqc_pca_3d(object, color_by, x_axis = "PC1", y_axis = "PC2", z_axis = "PC3")
}
\arguments{
\item{object}{tidymass-class object}

\item{color_by}{which column (sample_info) is used to color samples}

\item{x_axis}{Principal component for x axis, defalut is PC1}

\item{y_axis}{Principal component for y axis, defalut is PC2}

\item{z_axis}{Principal component for z axis, defalut is PC3}
}
\value{
A plotly object
}
\description{
interactive pca plot.
}
