% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_mass_dataset_fix.R
\name{merge_mass_dataset_fix}
\alias{merge_mass_dataset_fix}
\title{massdataset merge_mass_dataset fix bugs}
\usage{
merge_mass_dataset_fix(
  x,
  y,
  sample_direction = c("left", "right", "full", "inner"),
  variable_direction = c("left", "right", "full", "inner"),
  sample_by = c("sample_id"),
  variable_by = c("variable_id", "mz", "rt")
)
}
\arguments{
\item{x}{see merge_mass_dataset}

\item{y}{see merge_mass_dataset}

\item{sample_direction}{see merge_mass_dataset}

\item{variable_direction}{see merge_mass_dataset}

\item{sample_by}{see merge_mass_dataset}

\item{variable_by}{see merge_mass_dataset}
}
\value{
massdataset merge_mass_dataset fix bugs.
}
\description{
massdataset merge_mass_dataset fix bugs
}
\references{
massdataset::merge_mass_dataset

massdataset::merge_mass_dataset
}
