% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{metabolite_origin_upsetplot_fix}
\alias{metabolite_origin_upsetplot_fix}
\title{upsetplot for metabolite origin}
\usage{
metabolite_origin_upsetplot_fix(object, min_size = 1, counts = TRUE)
}
\arguments{
\item{object}{A mass_dataset object}

\item{min_size}{see ComplexUpset}

\item{counts}{see ComplexUpset}
}
\description{
upsetplot for metabolite origin
}
