% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms2_plot_mass_dataset_mz.R
\name{ms2_plot_mass_dataset_mz}
\alias{ms2_plot_mass_dataset_mz}
\title{Mirror plot between feature MS2 spectra and MS2 database spectra}
\usage{
ms2_plot_mass_dataset_mz(
  object,
  variable_id,
  variable_index,
  polarity = c("positive", "negative"),
  ms1.match.ppm = 25,
  ms2.match.ppm = 30,
  mz.ppm.thr = 400,
  database,
  interactive_plot = FALSE,
  show_mz = TRUE,
  show_detail = TRUE
)
}
\arguments{
\item{object}{see ms2_plot_mass_dataset.}

\item{variable_id}{see ms2_plot_mass_dataset.}

\item{variable_index}{see ms2_plot_mass_dataset.}

\item{polarity}{see ms2_plot_mass_dataset.}

\item{ms1.match.ppm}{see ms2_plot_mass_dataset.}

\item{ms2.match.ppm}{see ms2_plot_mass_dataset.}

\item{mz.ppm.thr}{see ms2_plot_mass_dataset.}

\item{database}{see ms2_plot_mass_dataset.}

\item{interactive_plot}{see ms2_plot_mass_dataset.}

\item{show_mz}{show m/z on plot.}

\item{show_detail}{show match details on plot.}
}
\value{
A ggplot object of mirror plot.
}
\description{
Mirror plot between feature MS2 spectra and MS2 database spectra
}
\references{
metid::ms2_plot_mass_dataset metminer::ms2_plot_mass_dataset_mz
}
