% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renderDataTable_formated.R
\name{renderDataTable_formated}
\alias{renderDataTable_formated}
\title{Render a Formatted DataTable with Extended Features}
\usage{
renderDataTable_formated(
  actions = NA,
  condition1 = NA,
  condition2 = NA,
  condition3 = NA,
  condition4 = NA,
  filename.a = "full_page",
  tbl
)
}
\arguments{
\item{actions}{Initial actions or JavaScript to be executed.}

\item{condition1}{First condition to check before rendering the table; if NULL, the table is not rendered.}

\item{condition2}{Second condition to check; similar to `condition1`.}

\item{condition3}{Third condition to check; similar to `condition1`.}

\item{condition4}{Fourth condition to check; similar to `condition1`.}

\item{filename.a}{The base filename for the CSV download button.}

\item{tbl}{The data object to be rendered as a DataTable.}
}
\value{
A DataTable object that is rendered within Shiny applications.
}
\description{
This function wraps `DT::renderDataTable` to provide additional functionality
including preset conditions, custom buttons for display control, and a CSV download
option. It allows for a highly customizable DataTable presentation with specific
interactive features tailored for enhanced user interaction.
}
\details{
The function integrates several customization options including custom buttons that
allow users to toggle the display length of the table and to download the full table data.
The `actions` parameter can be used to inject specific JavaScript actions directly into
the DataTable initialization.
}
