% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{summarize_massdataset}
\alias{summarize_massdataset}
\title{Summarize mass_dataset Object Information}
\usage{
summarize_massdataset(
  object,
  mode = c("positive", "negative"),
  show_processing = TRUE,
  show_qc = TRUE,
  color = TRUE
)
}
\arguments{
\item{object}{A S4 object of class `mass_dataset` containing metabolomics data}

\item{mode}{Ionization mode, either `"positive"` or `"negative"` (case-insensitive)}

\item{show_processing}{Logical indicating whether to display processing history.
Requires `object@process_info` slot. Default: `TRUE`}

\item{show_qc}{Logical indicating whether to perform QC sample detection.
Checks `sample_type` or `sample_id` columns. Default: `TRUE`}

\item{color}{Logical enabling ANSI color codes in output.
Set to `FALSE` for Shiny/RMarkdown. Default: `TRUE`}
}
\value{
A character string containing formatted summary report, suitable for
  rendering with `cat()` in Shiny `renderPrint` contexts
}
\description{
Generates a formatted report summarizing key metadata and processing history
of a `mass_dataset` object. Designed for integration with Shiny's `renderPrint`.
}
\examples{
\dontrun{
library(massdataset)


data(example_dataset)

# report in positive model
cat(summarize_massdataset(example_dataset, mode = "positive"))

# In shiny serve
output$summary <- renderPrint({
  cat(summarize_massdataset(object(), color = FALSE)
})
}

}
