% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{validate_file}
\alias{validate_file}
\title{validate mass_dataset file}
\usage{
validate_file(path, expected_polarity, object_label)
}
\arguments{
\item{path}{object file path}

\item{object_label}{name of uploaded object}

\item{expcted_polarity}{expected polarity of uploaded object.}
}
\value{
A character
}
\description{
validate mass_dataset file
}
