% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{validate_sample_files}
\alias{validate_sample_files}
\title{validate sample files}
\usage{
validate_sample_files(
  mode = "POS",
  QC_num.p,
  S_num.p,
  QC_num.n,
  S_num.n,
  sample_info_temp
)
}
\arguments{
\item{mode}{object file path}

\item{QC_num.p}{QC file number in positive model}

\item{S_num.p}{Subject file number in positive model}

\item{QC_num.n}{QC file number in negative model}

\item{S_num.n}{Subject file number in negative model}

\item{sample_info_temp}{sample information}
}
\value{
A character

Return a logical value indicating whether the input file matches the sample ID in the sample information.
}
\description{
validate sample files
}
